refactor: Interceptor, Resolver
