+++ src/main/java/gift/auth/Interceptor/AuthenticationInterceptor.java
@@ -1,4 +1,4 @@
-package gift.auth;
+package gift.auth.Interceptor;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
