+++ src/main/java/gift/auth/resolver/LoginMember.java
@@ -1,4 +1,4 @@
-package gift.auth;
+package gift.auth.resolver;
 
 import static java.lang.annotation.ElementType.PARAMETER;
 import static java.lang.annotation.RetentionPolicy.RUNTIME;
