+++ src/main/java/gift/product/application/ProductController.java
@@ -1,6 +1,6 @@
 package gift.product.application;
 
-import gift.auth.Authorized;
+import gift.auth.Interceptor.Authorized;
 import gift.member.domain.Role;
 import gift.product.application.dto.request.ProductCreateRequest;
 import gift.product.application.dto.request.ProductRequest;
